initramfs-dracut-in-docker: modified the recipe to inherit deploy.
It could alternatively be seen as a kind of an image, but we deliberately skip all kinds of interim steps (will be done with building initramfs natively), but there is really no need for complicating things. On the other hand, if it runs early enough, which it could, it is not guaranteed that the tmp/deploy/images direcory exists, and the recipe should not worry about it. Hence, it directly inherits deploy now.

Note though, that unless you still put
addtask deploy before do_build after do_compile
, it will not really put the file in the IMAGE_DEPLOY_DIR, when using
the EXTRA_IMAGEDEPENDS += "..." construct.

So avoidning doing this as an image type is not ideal. Note that with
the current version it has not been tested on a full build